miniContentPanel {miniUI}	R Documentation

Create a content panel

Description

Creates a panel for containing arbitrary content within a flex box container. This is mainly useful within miniPage or a miniTabPanel. You can use miniContentPanel to introduce padding and/or scrolling, but even if padding/scrolling aren't needed, it's a good idea to wrap your custom content into miniContentPanel as it fixes some odd behavior with percentage-based heights.

Usage

miniContentPanel(..., padding = 15, scrollable = TRUE)

Arguments

...
UI objects to be contained in the miniContentPanel. A single htmlwidget or plotOutput with height="100%" works well, as do fillRow/fillCol.

padding
Amount of padding to apply. Can be numeric (in pixels) or character (e.g. "3em").

scrollable
If TRUE, then content large enough to overflow the miniContentPanel will make scrollbars appear.

See Also

For more information, see the Designing Gadget UI article on shiny.rstudio.com.

Examples

library(shiny)

miniContentPanel(padding = 0,
  plotOutput("plot", height = "100%")
)

[Package miniUI version 0.1.1.1 Index]