miniTabstripPanel {miniUI}	R Documentation

Create a tabstrip panel

Description

miniTabstripPanel is a tabstrip panel that contains miniTabPanel elements. Similar to tabsetPanel, but optimized for small page sizes like mobile devices or the RStudio Viewer pane.

Usage

miniTabstripPanel(..., id = NULL, selected = NULL, between = NULL)

miniTabPanel(title, ..., value = title, icon = NULL)

Arguments

...
For miniTabstripPanel, miniTabPanel elements to include in the tabset. For miniTabPanel, UI elements to include within the tab.

id
If provided, you can use input$id in your server logic to determine which of the current tabs is active. The value will correspond to the value argument that is passed to miniTabPanel.

selected
The value (or, if none was supplied, the title) of the tab that should be selected by default. If NULL, the first tab will be selected.

between
A tag or list of tags that should be inserted between the content (above) and tabstrip (below).

title
Display title for tab.

value
The value that should be sent when miniTabstripPanel reports that this tab is selected. If omitted and miniTabstripPanel has an id, then the tab's title will be used as the value.

icon
Icon to appear on the tab; see icon.

See Also

For more information, see the Designing Gadget UI article on shiny.rstudio.com.

Examples

library(shiny)

miniTabstripPanel(
  miniTabPanel("Data", icon = icon("table"),
    selectInput("dataset", "Data set", ls("package:datasets"))),
  miniTabPanel("Subset", icon = icon("sliders"),
    uiOutput("subset_ui")
  )
)

[Package miniUI version 0.1.1.1 Index]